updated Users.js model to match database fields.
