+++ server/src/models/Users.js
@@ -1,62 +1,76 @@
 import { DataTypes, Model } from "sequelize";
 import sequelize from "../config/connection.js";
+import bcrypt from 'bcrypt';
 
 class User extends Model {}
 
 User.init(
-    {
-        id: {
-        type: DataTypes.INTEGER,
-        primaryKey: true,
-        autoIncrement: true,
-        },
-        name: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        unique: true,
-        },
-        address1: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        },
-        address2: {
-        type: DataTypes.STRING,
-        allowNull: true,
-        },
-        city: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        },
-        state: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        },
-        zipcode: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        },
-        phoneNumber: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        },
-        email: {
-        type: DataTypes.STRING,
-        allowNull: false,
-        unique: true,
-        },
-        password: {
-        type: DataTypes.STRING,
-        allowNull: true,
-        },
-        isAdmin: {
-        type: DataTypes.BOOLEAN,
-        defaultValue: false,
-        },
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
     },
-    {
-        tableName: "users",
-        sequelize,
-    }
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    address1: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    address2: {
+      type: DataTypes.STRING,
+      allowNull: true,
+    },
+    city: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    state: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    zip: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+    },
+    // phoneNumber: {
+    //   type: DataTypes.STRING,
+    //   allowNull: false,
+    // },
+    email: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    password: {
+      type: DataTypes.STRING,
+      allowNull: true,
+      defaultValue: null,
+    },
+    isAdmin: {
+      type: DataTypes.BOOLEAN,
+      defaultValue: false,
+    },
+  },
+  {
+    tableName: "users",
+    sequelize,
+    hooks: {
+      async beforeCreate(user) {
+        if (user.password) {
+          const salt = await bcrypt.genSalt(10);
+          user.password = await bcrypt.hash(user.password, salt);
+        }
+      },
+      async beforeUpdate(user) {
+        if (user.password) {
+          const salt = await bcrypt.genSalt(10);
+          user.password = await bcrypt.hash(user.password, salt);
+        }
+      },
+    },
+  }
 );
 
 export default User;
